+++ backend/app/controllers/Backend/SuggestController.php
@@ -24,27 +24,6 @@ class SuggestController extends BaseController {
     }
 
     /**
-     * Frontend new Suggest
-     *
-     * @return Redirect
-     */
-    public function newSuggest()
-    {
-        extract(Input::all());
-
-        $user = new SuggestModel(array(
-            'name' => $name,
-            'email' => $email,
-            'suggest' => $suggest
-        ));
-
-        $user->save();
-
-        return Redirect::back()->with('thank','感谢您的建议反馈，我们将做得更好');
-
-    }
-
-    /**
      * Backend Suggest delete
      *
      * @return Response
